+++ test/snapshots/classes/waybill.snapshot.test.ts
@@ -1,0 +1,10 @@
+import { WaybillSchema, serializeWaybill } from '../../../src/classes/waybill/index.js'
+import { createWaybill } from '../../factories/waybill.js'
+import { snapshotHarness } from './_harness.js'
+
+snapshotHarness({
+  className: 'Waybill',
+  schema: WaybillSchema,
+  serialize: serializeWaybill,
+  factory: createWaybill,
+})